backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.10.1
python-dotenv==1.1.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
httpx==0.28.1
nltk==3.9.1
yfinance==0.2.65
edgartools>=4.0.0
